requirements.txt
folium
streamlit_folium
numpy
pandas
pillow